+++ includes/admin/admin-functions.php
@@ -155,19 +155,19 @@ function bfg_remove_files_on_upgrade() {
  */
 add_filter( 'secure_signon_cookie', '__return_true' );
 
-/**
+/*
  * Prevent login with username (email only).
  *
  * @since 20180604
  */
 // remove_filter( 'authenticate', 'wp_authenticate_username_password', 20 );
 
-/**
+/*
  * Prevent non-SSL HTTP origins.
  *
  * @since 20180604
  */
-// add_filter( 'allowed_http_origins', 'bfg_allowed_http_origins' );
+add_filter( 'allowed_http_origins', 'bfg_allowed_http_origins' );
 function bfg_allowed_http_origins($allowed_origins) {
 
 	$whitelisted_origins = array();